added lots of gui stuff

+++ sim.h
@@ -2,7 +2,7 @@
 
 //--- SIMULATION PARAMETERS ------------------------------------------------------------------------
 const int DIM = 200;				//size of simulation grid
-double dt = 0.5;				//simulation time step
+float dt = 0.5;				//simulation time step
 float visc = 0.0001;				//fluid viscosity
 fftw_real *vx, *vy;             //(vx,vy)   = velocity field at the current moment
 fftw_real *vx0, *vy0;           //(vx0,vy0) = velocity field at the previous moment
